renamed json parser, start test parse errors
