+++ include/matador/object/json_object_mapper.hpp
@@ -16,15 +16,11 @@
 namespace matador {
 
 template < class T >
-class json_object_mapper : public generic_json_parser_ng<json_object_mapper<T>>
+class json_object_mapper : public generic_json_parser<json_object_mapper<T>>
 {
 public:
-  json_object_mapper() : generic_json_parser_ng<json_object_mapper<T>>(this) {}
+  json_object_mapper() : generic_json_parser<json_object_mapper<T>>(this) {}
 
-//  json_object_mapper(const char *str, T *value)
-//    : generic_json_parser_ng<json_object_mapper<T>>(this)
-//    ,
-//  {}
   T* from_string(const char *str);
 
   /// @cond OOS_DEV //
@@ -36,7 +32,7 @@ public:
   void on_end_array();
 
   void on_string(const std::string &value);
-  void on_number(typename generic_json_parser_ng<json_object_mapper<T>>::number_t value);
+  void on_number(typename generic_json_parser<json_object_mapper<T>>::number_t value);
   void on_bool(bool value);
   void on_null();
   /// @endcond OOS_DEV //
@@ -127,7 +123,7 @@ void json_object_mapper<T>::on_string(const std::string &value)
 }
 
 template<class T>
-void json_object_mapper<T>::on_number(typename generic_json_parser_ng<json_object_mapper<T>>::number_t value)
+void json_object_mapper<T>::on_number(typename generic_json_parser<json_object_mapper<T>>::number_t value)
 {
   if (object_key_.empty()) {
     if (value.is_real) {
