+++ src/utils/json_parser.cpp
@@ -9,7 +9,7 @@
 namespace matador
 {
 json_parser::json_parser()
-  : generic_json_parser_ng<json_parser>(this)
+  : generic_json_parser<json_parser>(this)
 {}
 
 //json json_parser::parse(std::istream &in)
@@ -100,7 +100,7 @@ void json_parser::on_string(const std::string &value)
   }
 }
 
-void json_parser::on_number(generic_json_parser_ng<json_parser>::number_t value)
+void json_parser::on_number(generic_json_parser<json_parser>::number_t value)
 {
   if (state_stack_.top()->is_object()) {
     if (value.is_real) {
